Moved and marked test.
